cleanup vestigial numba vs. cython code structure
