--- py/redrock/_rebin.py
@@ -1,60 +1,0 @@
-"""
-Low-level optimized functions for rebinning templates.
-"""
-
-from __future__ import absolute_import, division, print_function
-
-import numba
-
-# This code is purposely written in a very "C-like" way.  The logic
-# being that it may help numba optimization and also makes it easier
-# if it ever needs to be ported to Cython.  Actually Cython versions
-# of this code have already been tested and shown to perform no better
-# than numba on Intel haswell and KNL architectures.
-
-@numba.jit
-def _trapz_rebin(x, y, edges, results):
-    nbin = len(edges) - 1
-    nx = len(x)
-    i = 0  #- index counter for output
-    j = 0  #- index counter for inputs
-    yedge = 0.0
-    area = 0.0
-
-    while i < nbin:
-        #- Seek next sample beyond bin edge
-        while x[j] <= edges[i]:
-            j += 1
-
-        #- What is the y value where the interpolation crossed the edge?
-        yedge = y[j-1] + (edges[i]-x[j-1]) * (y[j]-y[j-1]) / (x[j]-x[j-1])
-
-        #- Is this sample inside this bin?
-        if x[j] < edges[i+1]:
-            area = 0.5 * (y[j] + yedge) * (x[j] - edges[i])
-            results[i] += area
-
-            #- Continue with interior bins
-            while x[j+1] < edges[i+1]:
-                j += 1
-                area = 0.5 * (y[j] + y[j-1]) * (x[j] - x[j-1])
-                results[i] += area
-
-            #- Next sample will be outside this bin; handle upper edge
-            yedge = y[j] + (edges[i+1]-x[j]) * (y[j+1]-y[j]) / (x[j+1]-x[j])
-            area = 0.5 * (yedge + y[j]) * (edges[i+1] - x[j])
-            results[i] += area
-
-        #- Otherwise the samples span over this bin
-        else:
-            ylo = y[j] + (edges[i]-x[j]) * (y[j] - y[j-1]) / (x[j] - x[j-1])
-            yhi = y[j] + (edges[i+1]-x[j]) * (y[j] - y[j-1]) / (x[j] - x[j-1])
-            area = 0.5 * (ylo+yhi) * (edges[i+1]-edges[i])
-            results[i] += area
-
-        i += 1
-
-    for i in range(nbin):
-        results[i] /= edges[i+1] - edges[i]
-
-    return
